This time the rightfull checkin of:
"Quite some refactoring (triggered by a bug that needed a fix) of the
integration code between SilvaDocument and SilvaExternalSources.

Added externalsource.py and _externalsource.py (mind the underscore).
The first will, depending on the availability of SilvaExternalSources,
import functions from the second.

The EditorSupport service provides access to this functionality."